feat: Separate board creation logic to boardRow class
