--- src/main/kotlin/minesweeper/MineBoardRow.kt
@@ -1,3 +1,0 @@
-package minesweeper
-
-class MineBoardRow(val mineCells: Array<Cell>)
